+++ lib/everything.rb
@@ -1,6 +1,7 @@
 require 'dotenv'
 Dotenv.load
 
+require 'pathname'
 require 'fastenv'
 require 'everything/version'
 require 'everything/logger'
@@ -8,7 +9,7 @@ require 'everything/piece'
 
 module Everything
   def self.path
-    Fastenv.everything_path
+    Pathname.new(Fastenv.everything_path)
   end
 end