+++ lib/everything/piece/content.rb
@@ -7,11 +7,26 @@ module Everything
         @piece_path = piece_path
       end
 
+      def absolute_dir
+        @absolute_dir ||= Everything.path.join(dir)
+      end
+
+      def absolute_path
+        @absolute_path ||= absolute_dir.join(file_name)
+      end
+
+      def dir
+        @dir ||= calculated_dir
+      end
+
       def file_name
         'index.md'
       end
 
       def file_path
+        # TODO: Could try a deprecation approach like http://seejohncode.com/2012/01/09/deprecating-methods-in-ruby/
+        deprecation_message = "Piece Content's #file_path is deprecated and will be removed soon. Use #absolute_path instead."
+        warn deprecation_message
         @file_path ||= File.join(piece_path, file_name)
       end
 
@@ -23,8 +38,12 @@ module Everything
         partitioned_text.last
       end
 
+      def path
+        @path ||= dir.join(file_name)
+      end
+
       def raw_markdown
-        @raw_markdown ||= File.read(file_path)
+        @raw_markdown ||= absolute_path.read
       end
 
       def raw_markdown=(value)
@@ -34,7 +53,7 @@ module Everything
       def save
         FileUtils.mkdir_p(piece_path)
 
-        File.write(file_path, @raw_markdown)
+        absolute_path.write(@raw_markdown)
       end
 
     private
@@ -43,6 +62,12 @@ module Everything
       def partitioned_text
         @partitioned_text ||= raw_markdown.partition("\n\n")
       end
+
+      def calculated_dir
+        full_pathname = Pathname.new(piece_path)
+        _relative_pathname = full_pathname.relative_path_from(Everything.path)
+      end
     end
   end
 end
+