+++ lib/everything/piece/metadata.rb
@@ -15,12 +15,31 @@ module Everything
         @piece_path = piece_path
       end
 
+      def absolute_dir
+        @absolute_dir ||= Everything.path.join(dir)
+      end
+
+      def absolute_path
+        @absolute_path ||= absolute_dir.join(file_name)
+      end
+
+      def dir
+        @dir ||= calculated_dir
+      end
+
       def file_path
+        # TODO: Could try a deprecation approach like http://seejohncode.com/2012/01/09/deprecating-methods-in-ruby/
+        deprecation_message = "Piece Metadata's #file_path is deprecated and will be removed soon. Use #absolute_path instead."
+        warn deprecation_message
         @file_path ||= File.join(piece_path, file_name)
       end
 
+      def path
+        @path ||= dir.join(file_name)
+      end
+
       def raw_yaml
-        @raw_yaml ||= YAML.load_file(file_path)
+        @raw_yaml ||= YAML.load_file(absolute_path)
       end
 
       def raw_yaml=(value)
@@ -30,7 +49,7 @@ module Everything
       def save
         FileUtils.mkdir_p(piece_path)
 
-        File.write(file_path, @raw_yaml)
+        absolute_path.write(@raw_yaml)
       end
 
       def_delegators :raw_yaml, :[]
@@ -40,6 +59,11 @@ module Everything
 
       def file_name
         'index.yaml'
+      end
+
+      def calculated_dir
+        full_pathname = Pathname.new(piece_path)
+        _relative_pathname = full_pathname.relative_path_from(Everything.path)
       end
     end
   end