+++ spec/everything/piece/content_spec.rb
@@ -1,35 +1,62 @@
+require 'fakefs/spec_helpers'
 require './spec/support/pieces'
 
 describe Everything::Piece::Content do
-  shared_context 'with tmp piece markdown file on disk' do
-    let(:tmp_piece_markdown_path) do
-      File.join(tmp_piece_path, 'index.md')
+  include_context 'with fake piece'
+
+  let(:content) do
+    described_class.new(fake_piece_path)
+  end
+
+  describe '#absolute_dir' do
+    let(:content_absolute_dir) do
+      Pathname.new('/fake/everything/path/grond-crawled-on')
     end
 
-    let(:given_markdown) do
-      <<MD
-# Piece Title Here
-
-The body is totally this right here.
-
-And it might even include multiple lines!
-MD
+    it "returns the content's absolute path" do
+      expect(content.absolute_dir).to eq(content_absolute_dir)
     end
 
-    before do
-      File.open(tmp_piece_markdown_path, 'w') do |markdown_file|
-        markdown_file.puts given_markdown
-      end
+    it 'memoizes the value' do
+      first_call_result = content.absolute_dir
+      second_call_result = content.absolute_dir
+      expect(first_call_result.object_id).to eq(second_call_result.object_id)
     end
   end
 
-  let(:content) do
-    described_class.new(tmp_piece_path)
+  describe '#absolute_path' do
+    let(:content_absolute_path) do
+      Pathname.new('/fake/everything/path/grond-crawled-on/index.md')
+    end
+
+    it "returns the content's absolute path" do
+      expect(content.absolute_path).to eq(content_absolute_path)
+    end
+
+    it 'memoizes the value' do
+      first_call_result = content.absolute_path
+      second_call_result = content.absolute_path
+      expect(first_call_result.object_id).to eq(second_call_result.object_id)
+    end
+  end
+
+  describe '#dir' do
+    let(:content_dir_relative_to_everything_path) do
+      Pathname.new('grond-crawled-on')
+    end
+
+    it "returns the content's path relative to everything path" do
+      expect(content.dir).to eq(content_dir_relative_to_everything_path)
+    end
+
+    it 'memoizes the value' do
+      first_dir_value = content.dir
+      second_dir_value = content.dir
+      expect(first_dir_value.object_id).to eq(second_dir_value.object_id)
+    end
   end
 
   describe '#file_name' do
-    include_context 'with tmp piece on disk'
-
     let(:expected_file_name) do
       'index.md'
     end
@@ -40,10 +67,8 @@ MD
   end
 
   describe '#file_path' do
-    include_context 'with tmp piece on disk'
-
     let(:expected_file_path) do
-      "#{tmp_piece_path}/index.md"
+      "#{fake_piece_path}/index.md"
     end
 
     it 'is the index.md under the piece' do
@@ -52,8 +77,7 @@ MD
   end
 
   describe '#title' do
-    include_context 'with tmp piece on disk'
-    include_context 'with tmp piece markdown file on disk'
+    include_context 'with fake piece content'
 
     let(:expected_title) do
       'Piece Title Here'
@@ -65,8 +89,7 @@ MD
   end
 
   describe '#body' do
-    include_context 'with tmp piece on disk'
-    include_context 'with tmp piece markdown file on disk'
+    include_context 'with fake piece content'
 
     let(:expected_body) do
     <<MD
@@ -81,9 +104,24 @@ MD
     end
   end
 
+  describe '#path' do
+    let(:content_path_relative_to_everything_path) do
+      Pathname.new('grond-crawled-on/index.md')
+    end
+
+    it "returns the content's path relative to everything path" do
+      expect(content.path).to eq(content_path_relative_to_everything_path)
+    end
+
+    it 'memoizes the value' do
+      first_path_value = content.path
+      second_path_value = content.path
+      expect(first_path_value.object_id).to eq(second_path_value.object_id)
+    end
+  end
+
   describe '#raw_markdown' do
-    include_context 'with tmp piece on disk'
-    include_context 'with tmp piece markdown file on disk'
+    include_context 'with fake piece content'
 
     let(:expected_raw_markdown) do
       given_markdown
@@ -94,21 +132,19 @@ MD
     end
 
     it 'memoizes the file read' do
-      allow(File)
+      allow(content.absolute_path)
         .to receive(:read)
         .and_call_original
 
       2.times { content.raw_markdown }
 
-      expect(File)
+      expect(content.absolute_path)
         .to have_received(:read)
         .exactly(:once)
     end
   end
 
   describe '#raw_markdown=' do
-    include_context 'with tmp piece on disk'
-
     let(:new_raw_markdown) do
       <<MD
 # New Markdown
@@ -125,78 +161,75 @@ MD
   end
 
   describe '#save' do
-    let(:tmp_dir) do
-      Dir.tmpdir
-    end
-
-    let(:tmp_piece_path) do
-      File.join(tmp_dir, 'a-test-piece')
-    end
-
     before do
+      FakeFS.activate!
+
       content.raw_markdown = "# Ship Shape"
     end
 
     after do
-      FileUtils.remove_entry(tmp_piece_path)
+      FileUtils.rm_rf(fake_piece_path)
+
+      FakeFS.deactivate!
     end
 
     context 'when the piece directory does not exist' do
+      before do
+        FileUtils.rm_rf(fake_piece_path)
+      end
+
       it 'creates the folder' do
-        expect(Dir.exist?(tmp_piece_path)).to eq(false)
+        expect(fake_piece_path).not_to exist
 
         content.save
 
-        expect(Dir.exist?(tmp_piece_path)).to eq(true)
+        expect(fake_piece_path).to exist
       end
 
       it 'creates the content markdown file' do
-        expect(File.exist?(content.file_path)).to eq(false)
+        expect(content.absolute_path).not_to exist
 
         content.save
 
-        expect(File.exist?(content.file_path)).to eq(true)
+        expect(content.absolute_path).to exist
       end
 
       it 'writes the markdown to the file' do
         content.save
 
-        expect(File.read(content.file_path)).to eq('# Ship Shape')
+        expect(content.absolute_path.read).to eq('# Ship Shape')
       end
     end
 
     context 'when the piece directory exists' do
-      before do
-        FileUtils.mkdir_p(tmp_piece_path)
-      end
-
       context 'when the content file does not exist' do
         it 'creates the content markdown file' do
-          expect(File.exist?(content.file_path)).to eq(false)
+          expect(content.absolute_path).not_to exist
 
           content.save
 
-          expect(File.exist?(content.file_path)).to eq(true)
+          expect(content.absolute_path).to exist
         end
 
         it 'writes the markdown to the file' do
           content.save
 
-          expect(File.read(content.file_path)).to eq('# Ship Shape')
+          expect(content.absolute_path.read).to eq('# Ship Shape')
         end
       end
 
       context 'when the content file already exists' do
         before do
-          File.write(content.file_path, '# Rolly Polly')
+          content.absolute_path.write('# Rolly Polly')
         end
 
         it 'overwrites the file with the correct markdown' do
           content.save
 
-          expect(File.read(content.file_path)).to eq('# Ship Shape')
+          expect(content.absolute_path.read).to eq('# Ship Shape')
         end
       end
     end
   end
 end
+