+++ spec/everything_spec.rb
@@ -4,16 +4,18 @@ describe Everything do
   end
 
   describe '.path' do
+    include_context 'with fake everything path env var'
+
     let(:expected_path) do
-      '/some/path/to/your/everything/repo/'
+      fake_everything_path
     end
 
-    before do
-      ENV['EVERYTHING_PATH'] = expected_path
+    it 'is a pathname' do
+      expect(described_class.path).to be_a_kind_of(Pathname)
     end
 
     it 'is the path from the environment' do
-      expect(described_class.path).to eq(expected_path)
+      expect(described_class.path.to_s).to eq(expected_path)
     end
   end
 end